API Documentation

Exported functions

DiffPointRasterisation.raster — Function

raster(grid_size, points, rotation, translation, [background, weight])

Interpolate points (multi-) linearly into an Nd-array of size grid_size.

Before points are interpolated into the array, each point $p$ is first transformed according to

\[\hat{p} = R p + t\]

with rotation $R$ and translation $t$.

Points $\hat{p}$ that fall into the N-dimensional hypercube with edges spanning from (-1, 1) in each dimension, are interpolated into the output array.

The total weight of each point is distributed onto the 2^N nearest pixels/voxels of the output array (according to the closeness of the voxel center to the coordinates of point $\hat{p}$) via N-linear interpolation.

DiffPointRasterisation.raster! — Method

raster!(out, points, rotation, translation, [background, weight])

Inplace version of raster.

Write output into out and return out.

DiffPointRasterisation.raster_project — Function

raster_project(grid_size, points, rotation, translation, [background, weight])

Interpolate N-dimensional points (multi-) linearly into an N-1 dimensional-array of size grid_size.

Before points are interpolated into the array, each point $p$ is first transformed according to

\[\hat{p} = P R p + t\]

with N-dimensional rotation $R$, projection $P$ and N-1-dmensional translation $t$. The projection simply drops the last coordinate of $R p$.

Points $\hat{p}$ that fall into the N-1-dimensional hypercube with edges spanning from (-1, 1) in each dimension, are interpolated into the output array.

The total weight of each point is distributed onto the 2^(N-1) nearest pixels/voxels of the output array (according to the closeness of the voxel center to the coordinates of point $\hat{p}$) via N-1-linear interpolation.

DiffPointRasterisation.raster_project! — Method

raster_project!(out, points, rotation, translation, [background, weight])

Inplace version of raster_project.

Write output into out and return out.

DiffPointRasterisation.raster_project_pullback! — Method

raster_project_pullback!(
    ds_dout, points, rotation, translation, [background, weight];
    [ds_dpoints, ds_drotation, ds_dtranslation, ds_dbackground, ds_dweight]
)

Pullback for raster_project(...)/raster_project!(...).

Take as input ds_dout the sensitivity of some quantity (s for "scalar") to the output out of the function raster_project(args...), as well as the exact same arguments args that were passed to raster_project, and return the sensitivities of s to the inputs args of the function raster_project()/raster_project!().

Optionally, pre-allocated output arrays for each input sensitivity can be specified as ds_d$INPUT_NAME, e.g. ds_dtranslation = [zeros(2) for _ in 1:8] for 3-dimensional points and a batch size of 8.

DiffPointRasterisation.raster_pullback! — Method

raster_pullback!(
    ds_dout, points, rotation, translation, [background, weight];
    [ds_dpoints, ds_drotation, ds_dtranslation, ds_dbackground, ds_dweight]
)

Pullback for raster(...)/raster!(...).

Take as input ds_dout the sensitivity of some quantity (s for "scalar") to the output out of the function raster(args...), as well as the exact same arguments args that were passed to raster, and return the sensitivities of s to the inputs args of the function raster()/raster!().

Optionally, pre-allocated output arrays for each input sensitivity can be specified as ds_d$INPUT_NAME, e.g. ds_dtranslation = [zeros(2) for _ in 1:8] for 2-dimensional points and a batch size of 8.

Private functions

DiffPointRasterisation.digitstuple — Method

digitstuple(k, Val(N))

Return a N-tuple containing the bit-representation of k

DiffPointRasterisation.voxel_shifts — Method

voxel_shifts(Val(N), [int_type])

Enumerate nearest neighbor coordinate shifts with respect to "upper left" voxel.

For a N-dimensional voxel grid, return a 2^N-tuple of N-tuples, where each element of the outer tuple is a cartesian coordinate shift from the "upper left" voxel.